+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 # Changelog
+## [1.3.11] - 2023-11-08
+### Breaking Changes
+- Removed api_key from estimate function on the sdk and added the same on the queryString, please refer examples/13-paymaster.ts for more info
+
+### Bug Fixes
+- Updated paymaster url to accept arka apiKey and chainId as queryString
+- Added optional parameters such as entryPointAddress and Factory walletAddress for custom chain interaction
+- Bug fixes for handling errors on connecting with custom chain interaction
+
+
 ## [1.3.10] - 2023-10-31
 ### New
 - Added getExchangeSupportedAssets to gets exchange supported tokens